+++ crates/cli/src/wasm_bindgen_test_runner.rs
@@ -411,12 +411,19 @@ fn rmain(cli: Cli) -> anyhow::Result<()> {
             }
             println!("Tests are now available at http://{addr}");
             thread::spawn(|| srv.run());
-            headless::run(&addr, &shell, driver_timeout, browser_timeout)?;
+            headless::run(
+                &addr,
+                &shell,
+                driver_timeout,
+                browser_timeout,
+                cli.nocapture,
+            )?;
         }
         TestMode::Browser { .. }
         | TestMode::DedicatedWorker { .. }
         | TestMode::SharedWorker { .. }
         | TestMode::ServiceWorker { .. } => {
+            let nocapture = cli.nocapture;
             let srv = server::spawn(
                 &if headless {
                     "127.0.0.1:0".parse().unwrap()
@@ -451,7 +458,7 @@ fn rmain(cli: Cli) -> anyhow::Result<()> {
             }
 
             thread::spawn(|| srv.run());
-            headless::run(&addr, &shell, driver_timeout, browser_timeout)?;
+            headless::run(&addr, &shell, driver_timeout, browser_timeout, nocapture)?;
         }
     }
     Ok(())